+++ kikuchipy/indexing/_static_pattern_matching.py
@@ -18,8 +18,7 @@
 
 from typing import Union, List
 
-import numpy as np
-from orix.crystal_map import CrystalMap
+from orix.crystal_map import CrystalMap, create_coordinate_arrays
 
 from kikuchipy.indexing import merge_crystal_maps, orientation_similarity_map
 from kikuchipy.indexing._pattern_matching import _pattern_match
@@ -55,7 +54,7 @@ class StaticPatternMatching:
         n_slices: int = 1,
         return_merged_crystal_map: bool = False,
         get_orientation_similarity_map: bool = False,
-    ) -> Union[CrystalMap, List[CrystalMap]]:
+    ) -> Union[CrystalMap, List[CrystalMap], None]:
         """Match each experimental pattern to all simulated patterns, of
         known crystal orientations in pre-computed dictionaries
         :cite:`chen2015dictionary,jackson2019dictionary`, to determine
@@ -66,9 +65,8 @@ class StaticPatternMatching:
         default, but a valid user-defined similarity metric may be used
         instead.
 
-        :class:`~orix.crystal_map.crystal_map.CrystalMap`'s for each
-        dictionary with "scores" and "simulation_indices" as properties
-        are returned.
+        :class:`~orix.crystal_map.CrystalMap`'s for each dictionary with
+        "scores" and "simulation_indices" as properties are returned.
 
         Parameters
         ----------
@@ -93,9 +91,8 @@ class StaticPatternMatching:
 
         Returns
         -------
-        xmaps : :class:`~orix.crystal_map.crystal_map.CrystalMap` or \
-                list of \
-                :class:`~orix.crystal_map.crystal_map.CrystalMap`
+        xmaps : orix.crystal_map.CrystalMap or list of \
+                orix.crystal_map.CrystalMap
             A crystal map for each dictionary loaded and one merged map
             if `return_merged_crystal_map = True`.
 
@@ -129,25 +126,28 @@ class StaticPatternMatching:
         # return the metric if it is not
         metric = _SIMILARITY_METRICS.get(metric, metric)
 
-        axes_manager = signal.axes_manager
-        spatial_arrays = _get_spatial_arrays(
-            shape=axes_manager.navigation_shape, extent=axes_manager.navigation_extent
+        am = signal.axes_manager
+
+        step_sizes = tuple([i.scale for i in am.navigation_axes])
+        coordinate_arrays, _ = create_coordinate_arrays(
+            shape=am.navigation_shape, step_sizes=step_sizes
         )
-        n_nav_dims = axes_manager.navigation_dimension
+
+        n_nav_dims = am.navigation_dimension
         if n_nav_dims == 0:
             xmap_kwargs = dict()
         elif n_nav_dims == 1:
-            scan_unit = axes_manager.navigation_axes[0].units
-            xmap_kwargs = dict(x=spatial_arrays, scan_unit=scan_unit)
+            scan_unit = am.navigation_axes[0].units
+            xmap_kwargs = dict(x=coordinate_arrays["x"], scan_unit=scan_unit)
         else:  # 2d
-            scan_unit = axes_manager.navigation_axes[0].units
+            scan_unit = am.navigation_axes[0].units
             xmap_kwargs = dict(
-                x=spatial_arrays[0], y=spatial_arrays[1], scan_unit=scan_unit
+                x=coordinate_arrays["x"], y=coordinate_arrays["y"], scan_unit=scan_unit
             )
 
         keep_n = min([keep_n] + [d.xmap.size for d in self.dictionaries])
 
-        # Naively let dask compute them seperately, should try in the
+        # Naively let dask compute them separately, should try in the
         # future combined compute for better performance
         xmaps = []
         patterns = signal.data
@@ -184,17 +184,3 @@ class StaticPatternMatching:
             xmaps = xmaps[0]
 
         return xmaps
-
-
-def _get_spatial_arrays(shape: tuple, extent: tuple) -> Union[tuple, np.ndarray]:
-    n_nav_dims = len(shape)
-    if n_nav_dims == 0:
-        return ()
-    if n_nav_dims == 1:
-        x0, x1 = extent
-        return np.linspace(x0, x1, shape[0])
-    else:
-        x0, x1, y0, y1 = extent
-        x = np.tile(np.linspace(x0, x1, shape[0]), shape[1])
-        y = np.tile(np.linspace(y0, y1, shape[1]), shape[0])
-        return x, y